+++ src/070_refactoring/start/scripts/glsl/world.js
@@ -1,5 +1,3 @@
-import { viewport } from "../helper/viewport";
-
 import {
   WebGLRenderer,
   Scene,
@@ -12,6 +10,8 @@ import {
   Vector2,
 } from "three";
 
+import { lerp, getWorldPosition } from "../helper/utils";
+
 const world = {
   init,
   render,
@@ -22,7 +22,9 @@ const os = [];
 const raycaster = new Raycaster();
 const pointer = new Vector2();
 
-function init() {
+function init(canvasRect, viewport) {
+  //   initScroller();
+  bindResizeEvents();
   world.renderer = new WebGLRenderer({
     canvas,
     antialias: true,
@@ -39,8 +41,12 @@ function init() {
     viewport.near,
     viewport.far
   );
-  world.camera.position.z = cameraZ;
+  world.camera.position.z = viewport.cameraZ;
 
+  initObjects(canvasRect);
+}
+
+function initObjects(canvasRect) {
   const els = document.querySelectorAll("[data-webgl]");
   els.forEach((el) => {
     const rect = el.getBoundingClientRect();
@@ -128,6 +134,94 @@ function raycast() {
 
     uHover.value = lerp(uHover.value, uHover.__endValue, 0.1);
   }
+}
+
+// function getWorldPosition(rect, canvasRect) {
+//   const x = rect.left + rect.width / 2 - canvasRect.width / 2;
+//   const y = -rect.top - rect.height / 2 + canvasRect.height / 2;
+//   return { x, y };
+// }
+
+function scroll(o) {
+  const newCanvasRect = canvas.getBoundingClientRect();
+  const {
+    $: { el },
+    mesh,
+  } = o;
+  const rect = el.getBoundingClientRect();
+  if (newCanvasRect) {
+    const { x, y } = getWorldPosition(rect, newCanvasRect);
+    // mesh.position.x = x;
+    mesh.position.y = y;
+  }
+  if (!newCanvasRect && canvasRect) {
+    const { x, y } = getWorldPosition(rect, canvasRect);
+    // mesh.position.x = x;
+    mesh.position.y = y;
+  }
+}
+
+function resize(o, newCanvasRect) {
+  const {
+    $: { el },
+    mesh,
+    geometry,
+    rect,
+  } = o;
+  const nextRect = el.getBoundingClientRect();
+  const { x, y } = getWorldPosition(nextRect, newCanvasRect);
+  mesh.position.x = x;
+  mesh.position.y = y;
+
+  // 大きさの変更
+  geometry.scale(nextRect.width / rect.width, nextRect.height / rect.height, 1);
+
+  o.rect = nextRect;
+}
+
+function onPointerMove(event) {
+  // calculate pointer position in normalized device coordinates
+  // (-1 to +1) for both components
+
+  pointer.x = (event.clientX / window.innerWidth) * 2 - 1;
+  pointer.y = -(event.clientY / window.innerHeight) * 2 + 1;
+}
+window.addEventListener("pointermove", onPointerMove);
+
+
+function bindResizeEvents() {
+  let timerId = null;
+
+  window.addEventListener("resize", () => {
+    clearTimeout(timerId);
+    timerId = setTimeout(() => {
+      console.log("resize");
+
+      const newCanvasRect = canvas.getBoundingClientRect();
+
+      // canvasサイズの変更
+      world.renderer.setSize(newCanvasRect.width, newCanvasRect.height, false);
+
+      // meshの位置とサイズの変更
+      os.forEach((o) => resize(o, newCanvasRect));
+
+      // cameraのProjectionMatrixの変更
+      const cameraWidth = newCanvasRect.width;
+      const cameraHeight = newCanvasRect.height;
+      const near = 1500;
+      const far = 4000;
+      const aspect = cameraWidth / cameraHeight;
+      const cameraZ = 2000;
+      const radian = 2 * Math.atan(cameraHeight / 2 / cameraZ);
+      const fov = radian * (180 / Math.PI);
+
+      world.camera.fov = fov;
+      world.camera.near = near;
+      world.camera.far = far;
+      world.camera.aspect = aspect;
+      world.camera.updateProjectionMatrix();
+    }, 500);
+  });
 }
 
 export default world;
